requirements.txt
pytorch-lightning
python-dotenv
hydra-core
hydra-colorlog
wandb
auraloss
datasets
pyloudnorm
einops
omegaconf
rich
plotly
librosa
eng-to-ipa
notebook

audio-diffusion-pytorch==0.0.43
audio-data-pytorch==0.0.16
